move get_keywords into its own file
